requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.12
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.59
botocore==1.35.59
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
elastic-transport==8.15.1
elasticsearch==8.16.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
Flask==2.1.3
Flask-Cors==5.0.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gotrue==2.11.3
groq==0.18.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-text-splitters==0.3.2
langsmith==0.1.143
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.54.4
opencv-python==4.10.0.84
orjson==3.10.11
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.1.0
postgrest==0.19.3
propcache==0.2.0
pydantic==2.9.2
pydantic-extra-types==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.19.1
pymongo==4.2.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
realtime==2.3.0
regex==2024.11.6
requests==2.32.3
requests-aws4auth==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
s3transfer==0.10.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.26.1
sentence-transformers==3.4.1
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.45.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.0
transformers==4.48.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
Werkzeug==2.0.3
wsproto==1.2.0
yarl==1.17.1
